ref(prism): move animation into movies pkg (#574)
