+++ src/prism/movies/doc.go
@@ -1,0 +1,2 @@
+// package movie contains animations used by prism views
+package movies
