+++ src/prism/movies/film-strips.go
@@ -1,4 +1,4 @@
-package traffic
+package movies
 
 import (
 	"strings"
