+++ src/prism/movies/spinners.go
@@ -1,4 +1,4 @@
-package traffic
+package movies
 
 import (
 	"sync"
