+++ src/prism/movies/unicode-spinners.go
@@ -1,4 +1,4 @@
-package traffic
+package movies
 
 var (
 	brailleWaveFrames = []string{
